Clock at Work
The program needs to memorize calendar date + hour and minute when you press 1, and the same thing when you press 2, but for example if you press 1 to 12:00 and 12:30 to memorize 2 2:00 and 12:30, but 24:00 is stored only 2 times. Not updating the time.
Finally, the program calculates the hours spent at work, but not calculated.